src/CloudScraper/requirements.txt
aiohttp
b2sdk
python-dotenv
pandas
cryptography
requests
backoff
python-binance
pyarrow
fastparquet